Extract method to process annotations data into new module
